requirements.txt
tensorflow
torch
Pillow
tqdm
safetensors
numpy
opencv-python
scipy